+++ pakyow-core/spec/unit/processes/proxy/server_spec.rb
@@ -1,13 +1,14 @@
 RSpec.describe Pakyow::Processes::Proxy::Server do
-  it "initializes with a port, host, and forwarded for value"
+  before do
+    allow(Async::HTTP::Client).to receive(:new)
+    allow(Async::HTTP::Endpoint).to receive(:parse)
+  end
 
-  describe "#call" do
-    it "passes the request to the client"
-    it "adds x-forwarded-to header to the request"
+  it "is deprecated" do
+    expect(Pakyow::Support::Deprecator.global).to receive(:deprecated).with(
+      described_class, { solution: "do not use" }
+    )
 
-    context "request fails" do
-      it "sleeps then retries"
-      it "retries for 15s"
-    end
+    described_class.new(port: 3000, host: "localhost", forwarded: nil)
   end
 end